chore: rename parent folder from "overview" -> "project"
