+++ ui/src/pages/deployment-details/deployment-details-form/section-general/section-general.tsx
@@ -14,7 +14,7 @@ import { Button, ButtonTheme } from 'design-system/components/button/button'
 import { ImageUpload } from 'design-system/components/image-upload/image-upload'
 import { InputContent } from 'design-system/components/input/input'
 import _ from 'lodash'
-import { EntitiesPicker } from 'pages/overview/entities/entities-picker'
+import { EntitiesPicker } from 'pages/project/entities/entities-picker'
 import { useContext } from 'react'
 import { useForm } from 'react-hook-form'
 import { FormContext } from 'utils/formContext/formContext'
