+++ ui/src/pages/deployment-details/deployment-details-form/section-source-images/section-source-images.tsx
@@ -13,8 +13,8 @@ import {
 import { Button } from 'design-system/components/button/button'
 import { InputContent, InputValue } from 'design-system/components/input/input'
 import _ from 'lodash'
-import { EntitiesPicker } from 'pages/overview/entities/entities-picker'
-import { ConnectionStatus } from 'pages/overview/storage/connection-status'
+import { EntitiesPicker } from 'pages/project/entities/entities-picker'
+import { ConnectionStatus } from 'pages/project/storage/connection-status'
 import { useContext, useState } from 'react'
 import { useForm } from 'react-hook-form'
 import { FormContext } from 'utils/formContext/formContext'
