+++ ui/src/pages/project/collections/collections.tsx
@@ -1,73 +1,70 @@
-import { ErrorState } from 'components/error-state/error-state'
 import { API_ROUTES } from 'data-services/constants'
-import { useProcessingServices } from 'data-services/hooks/processing-services/useProcessingServices'
+import { useCollections } from 'data-services/hooks/collections/useCollections'
 import { PageHeader } from 'design-system/components/page-header/page-header'
 import { PaginationBar } from 'design-system/components/pagination-bar/pagination-bar'
 import { Table } from 'design-system/components/table/table/table'
 import { TableSortSettings } from 'design-system/components/table/types'
-import { NewEntityDialog } from 'pages/overview/entities/new-entity-dialog'
-import { ProcessingServiceDetailsDialog } from 'pages/processing-service-details/processing-service-details-dialog'
+import { NewEntityDialog } from 'pages/project/entities/new-entity-dialog'
 import { useState } from 'react'
 import { useParams } from 'react-router-dom'
 import { STRING, translate } from 'utils/language'
 import { usePagination } from 'utils/usePagination'
 import { UserPermission } from 'utils/user/types'
-import { columns } from './processing-services-columns'
+import { columns } from './collection-columns'
 
-export const ProcessingServices = () => {
-  const { projectId, id } = useParams()
+export const Collections = () => {
+  const { projectId } = useParams()
   const [sort, setSort] = useState<TableSortSettings | undefined>({
     field: 'created_at',
     order: 'desc',
   })
   const { pagination, setPage } = usePagination()
-  const { items, userPermissions, total, isLoading, isFetching, error } =
-    useProcessingServices({
-      projectId,
-      pagination,
-      sort,
-    })
+  const { collections, userPermissions, total, isLoading, isFetching, error } =
+    useCollections(
+      {
+        projectId,
+        pagination,
+        sort,
+      },
+      true
+    )
   const canCreate = userPermissions?.includes(UserPermission.Create)
-
-  if (!isLoading && error) {
-    return <ErrorState error={error} />
-  }
 
   return (
     <>
       <PageHeader
-        title={translate(STRING.NAV_ITEM_PROCESSING_SERVICES)}
+        title={translate(STRING.NAV_ITEM_COLLECTIONS)}
         subTitle={translate(STRING.RESULTS, {
           total,
         })}
         isLoading={isLoading}
         isFetching={isFetching}
-        tooltip={translate(STRING.TOOLTIP_PROCESSING_SERVICE)}
+        tooltip={translate(STRING.TOOLTIP_COLLECTION)}
       >
         {canCreate && (
           <NewEntityDialog
-            collection={API_ROUTES.PROCESSING_SERVICES}
-            type="service"
+            collection={API_ROUTES.COLLECTIONS}
+            type="collection"
           />
         )}
       </PageHeader>
       <Table
-        items={items}
+        columns={columns(projectId as string)}
+        error={error}
         isLoading={isLoading}
-        columns={columns(projectId as string, canCreate)}
+        items={collections}
+        onSortSettingsChange={setSort}
         sortable
         sortSettings={sort}
-        onSortSettingsChange={setSort}
       />
-      {items?.length ? (
+      {collections?.length ? (
         <PaginationBar
           compact
           pagination={pagination}
-          total={total}
           setPage={setPage}
+          total={total}
         />
       ) : null}
-      {id ? <ProcessingServiceDetailsDialog id={id} /> : null}
     </>
   )
 }
