+++ ui/src/pages/project/collections/constants.tsx
