+++ ui/src/pages/project/entities/delete-entity-dialog.tsx
