+++ ui/src/pages/project/entities/details-form/collection-details-form.tsx
@@ -11,7 +11,7 @@ import { FormConfig } from 'components/form/types'
 import { Collection } from 'data-services/models/collection'
 import { Button, ButtonTheme } from 'design-system/components/button/button'
 import { IconType } from 'design-system/components/icon/icon'
-import { editableSamplingMethods } from 'pages/overview/collections/constants'
+import { editableSamplingMethods } from 'pages/project/collections/constants'
 import { useForm } from 'react-hook-form'
 import { STRING, translate } from 'utils/language'
 import { useFormError } from 'utils/useFormError'
