+++ ui/src/pages/project/entities/details-form/constants.ts
