+++ ui/src/pages/project/entities/details-form/entity-details-form.tsx
