+++ ui/src/pages/project/entities/details-form/processing-service-details-form.tsx
@@ -9,7 +9,7 @@ import { FormConfig } from 'components/form/types'
 import { ProcessingService } from 'data-services/models/processing-service'
 import { Button, ButtonTheme } from 'design-system/components/button/button'
 import { IconType } from 'design-system/components/icon/icon'
-import { ConnectionStatus } from 'pages/overview/processing-services/connection-status'
+import { ConnectionStatus } from 'pages/project/processing-services/connection-status'
 import { useForm } from 'react-hook-form'
 import { STRING, translate } from 'utils/language'
 import { useFormError } from 'utils/useFormError'
