+++ ui/src/pages/project/entities/details-form/types.ts
