+++ ui/src/pages/project/entities/devices.tsx
