+++ ui/src/pages/project/entities/entity-details-dialog.tsx
