+++ ui/src/pages/project/entities/styles.module.scss
