+++ ui/src/pages/project/processing-services/processing-services-columns.tsx
@@ -3,9 +3,9 @@ import { ProcessingService } from 'data-services/models/processing-service'
 import { BasicTableCell } from 'design-system/components/table/basic-table-cell/basic-table-cell'
 import { StatusTableCell } from 'design-system/components/table/status-table-cell/status-table-cell'
 import { CellTheme, TableColumn } from 'design-system/components/table/types'
-import { DeleteEntityDialog } from 'pages/overview/entities/delete-entity-dialog'
-import { UpdateEntityDialog } from 'pages/overview/entities/entity-details-dialog'
-import styles from 'pages/overview/entities/styles.module.scss'
+import { DeleteEntityDialog } from 'pages/project/entities/delete-entity-dialog'
+import { UpdateEntityDialog } from 'pages/project/entities/entity-details-dialog'
+import styles from 'pages/project/entities/styles.module.scss'
 import { Link } from 'react-router-dom'
 import { APP_ROUTES } from 'utils/constants'
 import { STRING, translate } from 'utils/language'
