+++ ui/src/pages/project/sidebar/useSidebarSections.tsx
