+++ ui/src/pages/project/storage/status-info/status-info.module.scss
