+++ ui/src/pages/project/summary/deployments-map.tsx
